refactor(platform): move icon mapping to shared lib
