+++ src/components/platform-navigation/PlatformIcon.tsx
@@ -1,7 +1,7 @@
 import type { PlatformName } from '@lynx-js/lynx-compat-data';
 import { withBase } from '@rspress/core/runtime';
 import { cn } from '../../lib/utils';
-import { mapPlatformNameToIconName } from '../api-table/compat-table/headers';
+import { mapPlatformNameToIconName } from '../../lib/platform';
 import { PlatformIconProps } from './types';
 import './icon.scss';
